requirements_lock.txt
acres==0.3.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
auto-py-to-exe==2.46.0
beautifulsoup4==4.13.4
bottle==0.13.2
bottle-websocket==0.2.9
browser-cookie3==0.20.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
colorclass==2.2.2
compressed-rtf==1.0.7
configobj==5.0.9
configparser==7.2.0
cryptography==44.0.2
distro==1.9.0
docstring_parser==0.16
dpcpp-cpp-rt==2025.0.5
easygui==0.98.3
easyocr==1.7.2
ebcdic==1.1.1
Eel==0.18.1
et_xmlfile==2.0.0
etelemetry==0.3.1
extract-msg==0.54.1
filelock==3.18.0
fitz==0.0.1.dev2
fsspec==2025.5.1
future==1.0.0
fuzzywuzzy==0.18.0
gevent==24.11.1
gevent-websocket==0.10.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.97.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-cloud-vision==3.10.2
google-crc32c==1.7.1
google-genai==1.20.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
intel-cmplr-lib-rt==2025.0.5
intel-cmplr-lib-ur==2025.0.5
intel-cmplr-lic-rt==2025.0.5
intel-opencl-rt==2025.0.5
intel-openmp==2025.0.5
intel-sycl-rt==2025.0.5
intel_extension_for_pytorch==2.7.10+xpu
isodate==0.6.1
jellyfish==1.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
keyboard==0.13.5
lark==1.1.9
lazy_loader==0.4
Levenshtein==0.27.1
looseversion==1.3.0
lxml==5.3.1
lz4==4.4.4
MarkupSafe==3.0.2
mkl==2025.0.1
mkl-dpcpp==2025.0.1
mpmath==1.3.0
msoffcrypto-tool==5.4.2
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
numpy==2.2.6
olefile==0.47
oletools==0.60.2
onemkl-sycl-blas==2025.0.1
onemkl-sycl-datafitting==2025.0.1
onemkl-sycl-dft==2025.0.1
onemkl-sycl-lapack==2025.0.1
onemkl-sycl-rng==2025.0.1
onemkl-sycl-sparse==2025.0.1
onemkl-sycl-stats==2025.0.1
onemkl-sycl-vm==2025.0.1
openai==1.82.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pcodedmp==1.2.6
pdf2image==1.17.0
pefile==2023.2.7
pillow==10.4.0
playwright==1.52.0
proto-plus==1.26.1
protobuf==6.31.1
prov==2.0.1
psutil==7.0.0
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.1
pydantic_core==2.33.0
pydot==3.0.4
pyee==13.0.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.2
PyMuPDF==1.26.0
pyparsing==3.2.3
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
pyxnat==1.6.3
PyYAML==6.0.2
RapidFuzz==3.13.0
rdflib==6.3.2
red-black-tree-mod==1.22
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
RTFDE==0.1.2
ruamel.yaml==0.18.12
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
segtok==1.5.11
selenium==4.33.0
setuptools==78.1.0
shadowcopy==0.0.4
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sympy==1.14.0
tabulate==0.9.0
tbb==2022.1.0
tcmlib==1.2.0
threadpoolctl==3.6.0
tifffile==2025.5.26
torch==2.7.0
torchaudio==2.7.0+cpu
torchvision==0.22.0
tqdm==4.67.1
traits==7.0.2
trio==0.30.0
trio-websocket==0.12.2
ttkbootstrap==1.13.8
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
umf==0.9.1
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
win_unicode_console==0.5
WMI==1.5.1
wsproto==1.2.0
yake==0.4.8
zope.event==5.0
zope.interface==7.2
